Update Task 2 from practicum 4
